requirements.txt
manga-ocr
pykakasi
opencv-python-headless
Pillow
requests
deep-translator
pyyaml
nodriver
easyocr
